Only allow users to enter sizes between 1 and 100
The program won't create a drawing area if the user enters a size less
than one. On the other hand, if they enter one greater than a hundred,
the drawing area will load slowly. This commit makes it so users can
only enter sizes greater than or equal to one and less than or equal to
one hundred. Doing so ensures that the drawing area will always load and
display correctly.